requirements.txt
torch
torchvision
torchaudio
# --- Core Server ---
fastapi
uvicorn[standard]
python-multipart
requests
aiofiles
jinja2
pandas
scikit-learn
autogluon.tabular

# --- AI & Vision Processing ---
transformers
accelerate
Pillow
packaging
av
# Use headless to avoid installing X11/GL libraries in Docker
opencv-python-headless
decord
imageio
numpy
einops

# --- Google Cloud & APIs ---
google-generativeai
google-cloud-aiplatform
google-genai
# CORRECTED: The ML Croissant library package name
mlcroissant

# --- Fine-Tuning (LoRA/QLoRA) ---
peft
# Bitsandbytes 0.44+ 
bitsandbytes
trl
datasets

# --- Audio ---
openai-whisper
yt-dlp
